chore: split up core/Mainchain.ts
