+++ core/src/BiddingCalculator.ts
@@ -2,7 +2,7 @@ import BigNumber from 'bignumber.js';
 import BiddingCalculatorData from './BiddingCalculatorData.js';
 import { bigIntMax, bigIntMin, bigNumberToBigInt } from './utils.js';
 import { BidAmountAdjustmentType, BidAmountFormulaType, type IBiddingRules } from './IBiddingRules.js';
-import { MICROGONS_PER_ARGON } from './Mainchain.js';
+import { MICROGONS_PER_ARGON } from './index.js';
 
 const MICRONOTS_PER_ARGONOT = MICROGONS_PER_ARGON;
