--- core/src/Mainchain.ts
@@ -1,415 +1,0 @@
-import BigNumber from 'bignumber.js';
-import { type ArgonClient, convertFixedU128ToBigNumber, MICROGONS_PER_ARGON } from '@argonprotocol/mainchain';
-import { bigIntMax, bigIntMin, bigNumberToBigInt } from './utils.js';
-import { type IWinningBid } from '@argonprotocol/commander-bot';
-import { MiningFrames, TICKS_PER_COHORT } from './MiningFrames.js';
-import type { ApiDecoration } from '@polkadot/api/types';
-import type { MainchainClients } from './MainchainClients.js';
-
-export type MainchainClient = ArgonClient;
-export { MICROGONS_PER_ARGON };
-export const BLOCK_REWARD_INCREASE_PER_INTERVAL = BigInt(1_000);
-export const BLOCK_REWARD_MAX = BigInt(5_000_000);
-export const BLOCK_REWARD_INTERVAL = 118;
-
-export class Mainchain {
-  public get archiveClientPromise(): Promise<MainchainClient> {
-    return this.clients.archiveClientPromise;
-  }
-
-  public get prunedClientOrArchivePromise(): Promise<MainchainClient> {
-    return this.clients.prunedClientPromise ?? this.clients.archiveClientPromise;
-  }
-
-  private cachedGenesisTick: number | null = null;
-
-  constructor(public clients: MainchainClients) {}
-
-  public async getClient(needsHistoricalBlocks: boolean): Promise<MainchainClient> {
-    return this.clients.get(needsHistoricalBlocks);
-  }
-
-  public async getMicronotsRequiredForBid(): Promise<bigint> {
-    const client = await this.prunedClientOrArchivePromise;
-    return await client.query.miningSlot.argonotsPerMiningSeat().then(x => x.toBigInt());
-  }
-
-  public async getFrameStartBlockNumbers(api: ApiDecoration<'promise'>): Promise<number[]> {
-    const frameStartBlockNumbers = await api.query.miningSlot.frameStartBlockNumbers();
-    return frameStartBlockNumbers.map(x => x.toNumber());
-  }
-
-  public async getLiquidityPoolPayout(): Promise<{ totalPoolRewards: bigint; totalActivatedCapital: bigint }> {
-    const client = await this.prunedClientOrArchivePromise;
-    const currentFrameId = await this.getCurrentFrameId();
-    const minersAtFrame = await client.query.miningSlot.minersByCohort(currentFrameId);
-    const vaultPools = await client.query.liquidityPools.vaultPoolsByFrame(currentFrameId);
-    let totalMicrogonsBid = 0n;
-    let totalActivatedCapital = 0n;
-    for (const miner of minersAtFrame) {
-      totalMicrogonsBid += miner.bid.toBigInt();
-    }
-    console.log('Liquidity bid pool at current frame', { totalMicrogonsBid, frameId: currentFrameId });
-    for (const [_vaultId, entry] of vaultPools) {
-      console.log('Capital activated', entry.toJSON());
-      for (const [_, balance] of entry.contributorBalances) {
-        totalActivatedCapital = balance.toBigInt();
-      }
-      if (!entry.distributedProfits.isSome) {
-        totalActivatedCapital -= entry.distributedProfits.value.toBigInt();
-      }
-    }
-
-    const totalPoolRewardsBn = BigNumber(totalMicrogonsBid).multipliedBy(0.8);
-    const totalPoolRewards = bigNumberToBigInt(totalPoolRewardsBn);
-
-    return {
-      totalPoolRewards, //: 100_000 * MICROGONS_PER_ARGON,
-      totalActivatedCapital, //: 998_000 * MICROGONS_PER_ARGON,
-    };
-  }
-
-  public async getCurrentTick(): Promise<number> {
-    const client = await this.prunedClientOrArchivePromise;
-    return (await client.query.ticks.currentTick()).toNumber();
-  }
-
-  public async getTicksSinceGenesis(currentTick: number): Promise<number> {
-    if (this.cachedGenesisTick === null) {
-      const client = await this.prunedClientOrArchivePromise;
-      this.cachedGenesisTick = (await client.query.ticks.genesisTick()).toNumber();
-    }
-    return currentTick - this.cachedGenesisTick;
-  }
-
-  public async getMicrogonsPerBlockForMiner(api: ApiDecoration<'promise'>) {
-    const minerPercent = convertFixedU128ToBigNumber(api.consts.blockRewards.minerPayoutPercent.toBigInt());
-    const microgonsPerBlock = await api.query.blockRewards.argonsPerBlock().then(x => x.toBigInt());
-    return bigNumberToBigInt(minerPercent.times(microgonsPerBlock));
-  }
-
-  public async minimumBlockRewardsAtTick(
-    currentTick: number,
-  ): Promise<{ rewardsPerBlock: bigint; amountToMinerPercent: BigNumber; ticksSinceGenesis: number }> {
-    const client = await this.prunedClientOrArchivePromise;
-    const ticksSinceGenesis = await this.getTicksSinceGenesis(currentTick);
-    const initialReward = 500_000n; // Initial microgons reward per block
-    const amountToMiner = convertFixedU128ToBigNumber(client.consts.blockRewards.minerPayoutPercent.toBigInt());
-
-    // Calculate the number of intervals
-    const numIntervals = Math.floor(ticksSinceGenesis / BLOCK_REWARD_INTERVAL);
-
-    // Calculate the current reward per block
-    const currentReward = initialReward + BigInt(numIntervals) * BLOCK_REWARD_INCREASE_PER_INTERVAL;
-    const reward = bigIntMin(currentReward, BLOCK_REWARD_MAX);
-    return { rewardsPerBlock: reward, amountToMinerPercent: amountToMiner, ticksSinceGenesis };
-  }
-
-  public async getMinimumMicronotsMinedDuringTickRange(tickStart: number, tickEnd: number): Promise<bigint> {
-    const client = await this.prunedClientOrArchivePromise;
-    const halvingStartTick = client.consts.blockRewards.halvingBeginTick.toNumber();
-    const halvingTicks = client.consts.blockRewards.halvingTicks.toNumber();
-    // eslint-disable-next-line prefer-const
-    let { rewardsPerBlock, amountToMinerPercent } = await this.minimumBlockRewardsAtTick(tickStart);
-    let totalRewards = 0n;
-    for (let i = tickStart; i < tickEnd; i++) {
-      const elapsedTicks = await this.getTicksSinceGenesis(i);
-      if (elapsedTicks >= halvingStartTick) {
-        const halvings = Math.floor((elapsedTicks - halvingStartTick) / halvingTicks);
-        rewardsPerBlock = BigInt(Number(BLOCK_REWARD_MAX) / (halvings + 1));
-      } else if (elapsedTicks % BLOCK_REWARD_INTERVAL === 0) {
-        rewardsPerBlock += BLOCK_REWARD_INCREASE_PER_INTERVAL;
-        rewardsPerBlock = bigIntMin(rewardsPerBlock, BLOCK_REWARD_MAX);
-      }
-      totalRewards += rewardsPerBlock;
-    }
-    return bigNumberToBigInt(amountToMinerPercent.times(totalRewards));
-  }
-
-  public async getCurrentArgonTargetPrice(): Promise<number> {
-    const client = await this.prunedClientOrArchivePromise;
-    const argonPrice = (await client.query.priceIndex.current()).unwrapOrDefault();
-    return convertFixedU128ToBigNumber(argonPrice.argonUsdTargetPrice.toBigInt()).toNumber();
-  }
-
-  public async getNextCohortSize(): Promise<number> {
-    const client = await this.prunedClientOrArchivePromise;
-    return (await client.query.miningSlot.nextCohortSize()).toNumber();
-  }
-
-  public async getRetiringCohortSize(): Promise<number> {
-    const client = await this.prunedClientOrArchivePromise;
-    const nextFrameId = await client.query.miningSlot.nextFrameId();
-    const rollingCohortId = nextFrameId.toNumber() - 10;
-    if (rollingCohortId < 1) return 0;
-    return await client.query.miningSlot.minersByCohort(rollingCohortId).then(x => x.length);
-  }
-
-  public async getActiveMinersCount(): Promise<number> {
-    const client = await this.prunedClientOrArchivePromise;
-    const activeMiners = (await client.query.miningSlot.activeMinersCount()).toNumber();
-    return Math.max(activeMiners, 100);
-  }
-
-  public async getAggregateBidCosts(): Promise<bigint> {
-    const client = await this.prunedClientOrArchivePromise;
-    const bidsPerFrame = await client.query.miningSlot.minersByCohort.entries();
-
-    let aggregateBidCosts = 0n;
-    for (const [_, cohortData] of bidsPerFrame) {
-      aggregateBidCosts += cohortData.reduce((acc, bid) => acc + bid.bid.toBigInt(), 0n);
-    }
-
-    return aggregateBidCosts;
-  }
-
-  public async getRecentSeatSummaries(): Promise<
-    { biddingFrameId: number; seats: number; lowestWinningBid: bigint; highestWinningBid: bigint }[]
-  > {
-    const client = await this.prunedClientOrArchivePromise;
-    const bidsPerFrame = await client.query.miningSlot.minersByCohort.entries();
-
-    const summaries = [];
-    for (const [frameIdRaw, cohortData] of bidsPerFrame) {
-      const bids = cohortData.map(x => x.bid.toBigInt());
-      const lowestWinningBid = bigIntMin(...bids);
-      const highestWinningBid = bigIntMax(...bids);
-      summaries.push({
-        biddingFrameId: Number(frameIdRaw.toHuman()) - 1,
-        seats: cohortData.length,
-        lowestWinningBid,
-        highestWinningBid,
-      });
-    }
-
-    return summaries.sort((a, b) => b.biddingFrameId - a.biddingFrameId);
-  }
-
-  public async getAggregateBlockRewards(): Promise<{
-    microgons: bigint;
-    micronots: bigint;
-  }> {
-    const client = await this.prunedClientOrArchivePromise;
-    const blockRewards = await client.query.blockRewards.blockRewardsByCohort();
-    const nextCohortId = blockRewards.pop()?.[0].toNumber() ?? 1;
-    const currentCohortId = nextCohortId - 1;
-
-    const currentTick = await this.getCurrentTick();
-    const tickAtStartOfCurrentSlot = await this.getTickAtStartOfCurrentSlot();
-    const ticksElapsedToday = currentTick - tickAtStartOfCurrentSlot;
-
-    const rewards = { microgons: 0n, micronots: 0n };
-
-    for (const [cohortId, blockReward] of blockRewards) {
-      const fullRotationsSinceCohortStart = currentCohortId - cohortId.toNumber();
-      const ticksSinceCohortStart = fullRotationsSinceCohortStart * 1_440 + ticksElapsedToday;
-      const startingTick = currentTick - ticksSinceCohortStart;
-      const endingTick = startingTick + TICKS_PER_COHORT;
-      const microgonsMinedInCohort = (blockReward.toBigInt() * BigInt(TICKS_PER_COHORT)) / 10n;
-      const micronotsMinedInCohort =
-        (await this.getMinimumMicronotsMinedDuringTickRange(startingTick, endingTick)) / 10n;
-      rewards.microgons += microgonsMinedInCohort;
-      rewards.micronots += micronotsMinedInCohort;
-    }
-
-    return rewards;
-  }
-
-  public async fetchMicrogonsInCirculation(): Promise<bigint> {
-    const client = await this.prunedClientOrArchivePromise;
-    return (await client.query.balances.totalIssuance()).toBigInt();
-  }
-
-  public async fetchMicrogonsMinedPerBlockDuringNextCohort(): Promise<bigint> {
-    const client = await this.prunedClientOrArchivePromise;
-    return await client.query.blockRewards.argonsPerBlock().then(x => x.toBigInt());
-  }
-
-  public async fetchMicrogonExchangeRatesTo(api?: ApiDecoration<'promise'>): Promise<{
-    USD: bigint;
-    ARGNOT: bigint;
-    ARGN: bigint;
-    BTC: bigint;
-  }> {
-    api ??= await this.archiveClientPromise;
-    const microgonsForArgon = BigInt(1 * MICROGONS_PER_ARGON);
-    const priceIndexRaw = await api.query.priceIndex.current();
-    if (priceIndexRaw.isNone) {
-      return {
-        USD: microgonsForArgon,
-        ARGNOT: microgonsForArgon,
-        ARGN: microgonsForArgon,
-        BTC: microgonsForArgon,
-      };
-    }
-
-    const priceIndex = priceIndexRaw.value;
-    const usdForArgon = convertFixedU128ToBigNumber(priceIndex.argonUsdPrice.toBigInt());
-    const usdForArgnot = convertFixedU128ToBigNumber(priceIndex.argonotUsdPrice.toBigInt());
-    const usdForBtc = convertFixedU128ToBigNumber(priceIndex.btcUsdPrice.toBigInt());
-
-    // These exchange rates should be relative to the argon
-    const microgonsForUsd = this.calculateExchangeRateInMicrogons(BigNumber(1), usdForArgon);
-    const microgonsForArgnot = this.calculateExchangeRateInMicrogons(usdForArgnot, usdForArgon);
-    const microgonsForBtc = this.calculateExchangeRateInMicrogons(usdForBtc, usdForArgon);
-
-    return {
-      ARGN: microgonsForArgon,
-      USD: microgonsForUsd,
-      ARGNOT: microgonsForArgnot,
-      BTC: microgonsForBtc,
-    };
-  }
-
-  public async getNextSlotRange(): Promise<[number, number]> {
-    const client = await this.prunedClientOrArchivePromise;
-    const nextSlotRangeBytes = await client.rpc.state.call('MiningSlotApi_next_slot_era', '');
-    const nextSlotRangeRaw = client.createType('(u64, u64)', nextSlotRangeBytes);
-    return [nextSlotRangeRaw[0].toNumber(), nextSlotRangeRaw[1].toNumber()];
-  }
-
-  public async getTickAtStartOfNextCohort(): Promise<number> {
-    return (await this.getNextSlotRange())[0];
-  }
-
-  public async getTickAtStartOfAuctionClosing(): Promise<number> {
-    const client = await this.prunedClientOrArchivePromise;
-    const tickAtStartOfNextCohort = await this.getTickAtStartOfNextCohort();
-    const ticksBeforeBidEndForVrfClose = (
-      await client.query.miningSlot.miningConfig()
-    ).ticksBeforeBidEndForVrfClose.toNumber();
-    return tickAtStartOfNextCohort - ticksBeforeBidEndForVrfClose;
-  }
-
-  public async getTickAtStartOfCurrentSlot(): Promise<number> {
-    const tickAtStartOfNextCohort = await this.getTickAtStartOfNextCohort();
-    return tickAtStartOfNextCohort - 1_440;
-  }
-
-  public async fetchWinningBids(): Promise<(IWinningBid & { micronotsStakedPerSeat: bigint })[]> {
-    const client = await this.prunedClientOrArchivePromise;
-    const nextCohort = await client.query.miningSlot.bidsForNextSlotCohort();
-    return nextCohort.map((c, i): IWinningBid & { micronotsStakedPerSeat: bigint } => {
-      const address = c.accountId.toHuman();
-      const subAccountIndex = undefined;
-      const lastBidAtTick = c.bidAtTick.toNumber();
-      const bidPosition = i;
-      const microgonsPerSeat = c.bid.toBigInt();
-      const micronotsStakedPerSeat = c.argonots.toBigInt();
-      return { address, subAccountIndex, lastBidAtTick, bidPosition, microgonsPerSeat, micronotsStakedPerSeat };
-    });
-  }
-
-  public async fetchWinningBidAmountsForFrame(frameId: number): Promise<bigint[]> {
-    const client = await this.prunedClientOrArchivePromise;
-    const winningBids = await client.query.miningSlot.minersByCohort(frameId);
-    return winningBids.map(bid => bid.bid.toBigInt());
-  }
-
-  public async fetchPreviousDayWinningBidAmounts(): Promise<bigint[]> {
-    const startingFrameId = MiningFrames.calculateCurrentFrameIdFromSystemTime();
-    let frameIdToCheck = startingFrameId;
-    while (true) {
-      // We must loop backwards until we find a frame with winning bids
-      if (frameIdToCheck < startingFrameId - 10) {
-        // We've checked the last 10 frames and found no winning bids, so we're done
-        return [];
-      }
-      const winningBids = await this.fetchWinningBidAmountsForFrame(frameIdToCheck);
-      if (winningBids.length > 0) {
-        return winningBids;
-      }
-      frameIdToCheck--;
-    }
-  }
-
-  public async getCurrentFrameId(): Promise<number> {
-    const client = await this.prunedClientOrArchivePromise;
-    const nextFrameId = await client.query.miningSlot.nextFrameId();
-    return nextFrameId.toNumber() - 1; // Subtract 1 to get the current frame ID
-  }
-
-  public async forEachFrame<T>(
-    iterateByEpoch: boolean,
-    callback: (
-      justEndedFrameId: number,
-      api: MainchainClient,
-      meta: {
-        blockNumber: number;
-        specVersion: number;
-      },
-      abortController: AbortController,
-    ) => Promise<T>,
-  ): Promise<T[]> {
-    const client = await this.archiveClientPromise;
-    const abortController = new AbortController();
-    const seenFrames = new Set<number>();
-    const results: T[] = [];
-
-    let api = client as ApiDecoration<'promise'>;
-    let apiSpecVersion = api.runtimeVersion.specVersion.toNumber();
-    let frameStartBlockNumbers = await this.getFrameStartBlockNumbers(api);
-    let startingFrameId = await api.query.miningSlot.nextFrameId().then(x => x.toNumber() - 1);
-
-    for (let i = 0; i < frameStartBlockNumbers.length; i++) {
-      const currentFrameId = startingFrameId - i;
-      const blockNumber = frameStartBlockNumbers[i];
-      const hasAlreadySeenThisFrame = seenFrames.has(currentFrameId);
-      let shouldIterateThisFrame = !hasAlreadySeenThisFrame;
-      if (iterateByEpoch) {
-        shouldIterateThisFrame = i === 0 && !hasAlreadySeenThisFrame;
-      }
-
-      const iterateThisFrame = async () => {
-        console.log(`Exploring frame ${currentFrameId}`);
-        const result = await callback(
-          currentFrameId,
-          api as any,
-          { blockNumber, specVersion: apiSpecVersion },
-          abortController,
-        );
-        results.push(result);
-      };
-
-      if (shouldIterateThisFrame) {
-        await iterateThisFrame();
-        if (abortController.signal.aborted || currentFrameId <= 1) {
-          break; // Stop processing if the abort signal is triggered
-        }
-      }
-      seenFrames.add(currentFrameId);
-
-      const isLastFrame = i === frameStartBlockNumbers.length - 1;
-
-      if (isLastFrame) {
-        const nextBlockHash = await client.rpc.chain.getBlockHash(blockNumber - 1);
-        const nextApi = await client.at(nextBlockHash);
-        const nextApiSpecVersion = nextApi.runtimeVersion.specVersion.toNumber();
-        if (nextApiSpecVersion < 124) {
-          // frameStartBlockNumbers is not available in versions < 124
-          if (!shouldIterateThisFrame) {
-            await iterateThisFrame();
-          }
-          break;
-        }
-        api = nextApi;
-        apiSpecVersion = nextApiSpecVersion;
-        frameStartBlockNumbers = await this.getFrameStartBlockNumbers(api);
-        startingFrameId = currentFrameId - 1;
-        console.log(`Reloaded frame start block numbers for next frame ${startingFrameId}`);
-        i = -1;
-      }
-    }
-    return results;
-  }
-
-  private calculateExchangeRateInMicrogons(usdAmount: BigNumber, usdForArgon: BigNumber): bigint {
-    const oneArgonInMicrogons = BigInt(MICROGONS_PER_ARGON);
-    const usdAmountBn = BigNumber(usdAmount);
-    const usdForArgonBn = BigNumber(usdForArgon);
-    if (usdAmountBn.isZero() || usdForArgonBn.isZero()) return oneArgonInMicrogons;
-
-    const argonsRequired = usdAmountBn.dividedBy(usdForArgonBn);
-    return bigNumberToBigInt(argonsRequired.multipliedBy(MICROGONS_PER_ARGON));
-  }
-}
