+++ core/src/index.ts
@@ -1,11 +1,14 @@
 import BiddingCalculator from './BiddingCalculator.js';
 import createBidderParams, { Helper as BiddingParamsHelper } from './createBidderParams.js';
 import BiddingCalculatorData from './BiddingCalculatorData.js';
+export { type ArgonClient, MICROGONS_PER_ARGON } from '@argonprotocol/mainchain';
 
 export * from './IBiddingRules.js';
 export * from './IBidderParams.js';
-export * from './Mainchain.js';
+export * from './PriceIndex.js';
 export * from './MainchainClients.js';
+export * from './FrameIterator.js';
+export * from './Mining.js';
 export * from './MiningFrames.js';
 export * from './NetworkConfig.js';
 export * from './Accountset.js';
@@ -20,5 +23,4 @@ export * from './IBitcoinBlockMeta.js';
 
 export { convertBigIntStringToNumber, bigNumberToBigInt, JsonExt, filterUndefined, createNanoEvents } from './utils.js';
 
-export default BiddingCalculator;
 export { BiddingParamsHelper, createBidderParams, BiddingCalculatorData, BiddingCalculator };
