+++ src-vue/components/InputArgon.vue
@@ -21,8 +21,7 @@ import * as Vue from 'vue';
 import BigNumber from 'bignumber.js';
 import { useCurrency } from '../stores/currency';
 import InputNumber from './InputNumber.vue';
-import { bigNumberToBigInt } from '@argonprotocol/commander-core';
-import { MICROGONS_PER_ARGON } from '@argonprotocol/commander-core/src/Mainchain';
+import { bigNumberToBigInt, MICROGONS_PER_ARGON } from '@argonprotocol/commander-core';
 
 const currency = useCurrency();
